Added RMS calculation stuff, and PMP sending stuff

+++ Hardware/PIC32 Based Hardware/ADC Module/ADC Driver Code/SEAD_DAQ_v2.0c_Power/SB_Computation.h
@@ -10,7 +10,7 @@
 
 #include "DMA_Transfer.h"
 
-uint8_t SB_RMS(SampleBuffer *);
+uint32_t SB_RMS(SampleBuffer *);
 
 #endif	/* SB_COMPUTATION_H */
